reorganize dataset pacakge
