+++ yews/datasets/__init__.py
@@ -1,29 +1,11 @@
 from .base import BaseDataset
+from .base import Dataset
+from .base import DirDataset
+from .base import FileDataset
 from .base import is_dataset
 from .base import PathDataset
-from .dirs import DatasetArrayFolder
-from .dirs import DatasetFolder
-from .dirs import DirDataset
-from .files import DatasetArray
-from .files import FileDataset
 from .packaged_datasets import Mariana
 from .packaged_datasets import PackagedDataset
 from .packaged_datasets import SCSN
 from .packaged_datasets import Wenchuan
-from .sac import MarianaFromSource
 from .utils import *
-
-__all__ = (
-    'BaseDataset',
-    'PathDataset',
-    'FileDataset',
-    'DirDataset',
-    'DatasetArray',
-    'DatasetFolder',
-    'DatasetArrayFolder',
-    'PackagedDataset',
-    'Wenchuan',
-    'Mariana',
-    'SCSN',
-    'MarianaFromSource',
-)
