--- yews/utils/__init__.py
@@ -1,2 +1,0 @@
-from .io import *
-from .path import *
